third_party/WebKit/LayoutTests/platform/mac/external/wpt/web-animations/interfaces/AnimationTimeline/document-timeline-expected.txt

This is a testharness.js-based test.
PASS document.timeline identity tests 
FAIL document.timeline.currentTime value tests assert_equals: document.timeline.currentTime matches requestAnimationFrame time expected 176.35800000061863 but got 192.25933333291323
PASS document.timeline.currentTime liveness tests 
PASS iframe document.timeline.currentTime liveness tests 
PASS document.timeline.currentTime time should be the same for all RAF callbacks in a frame 
Harness: the test ran to completion.
